requirements.txt
streamlit==1.31.0
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.4.0
spacy==3.7.2
